fix: match prefix and not whole file name
- This was needed because some graphs added extra information after the
prefix.

- However this solution brings the problem of the numering of the file
being incremented by steps different than one. For example, a graph
script uses a prefix to generate 3 files. A second run should have a
prefix of the type prefix-1, instead it will have a prefix of the type
prefix-3. It is still possible to discern the order and to which run a
file corresponds to.