fix: session options is a map
- There was a specific case where if a workload had session options,
like most do, and after that a workload did not have session options,
like the case of the foreground threads, then on those workload the
session options would be defined and thus the worload would run on
the wrong configurations.

- A simple fix is just assume the session options is like a map, just
like workload_settings is. This way if session_options[workload] will be
undefined and the default configurations will be used.