feat: add size event context to energy report
- For now there has been a compromise between what is captured with the
erm library and what is possible to extract from the rocksdb logs.

- Parsing the rocksdb logs takes a substantial amount of time, however
making changes to the information we collect via erm library requires
running the tests again.

- This strategy is aiming at extracting the information we already have.
The optimal strategy would be to caputre the information via erm library
but not always do we know beforehand what type of data we will need.