feat: energy report parser and plotting utils
- It is possible to generate an Event from a report log json. With this
feature it is possible to only generate this log on the compute server.
Plot views will then be generated on a general purpose server. This
refactor was a must because plot generation was taking a considerable
chunck of time. Plotting increased test times in the order of magnitude of
hours—each test has around 20-30 runs and each run is repeated at least
two times—and made plotting on the congested and compute intensive server a nuisance.

- To facilitate plot generation a wrapper around a plot library was
created. Many of its features are just a one to one implementation of the
matplotlib API. There is, however, some value to this wrapper; the
plotting backend can change with minimal friction  and while developing
the wrapper I learn matplotlib in a slightly deeper level.