refactor: results structure and cleanup
- The previous structure did not anticipate the increase of log files
and artifacts produced in each repetitions. In fact in the beginning
repetitions were not envisioned. They were a necessity to check for
variance between runs. However, when added the repetition logic I was in
a hurry and did not make the best design choice.

- The new structure is organized by session name dirs. Each session name
dir is organized by timestamp session dir. The current session dir is
organized in deltas dirs. Each delta dir is organized in repetition
number.